refactor(extension): migrate to WXT + 5 fixes
- Replace @crxjs/vite-plugin (beta stagnant) with WXT 0.20
  → HMR dev mode, npm run zip for Chrome Web Store packaging
- Filter SYSTEM_OWNERS in background.ts — context menu no longer
  appears on github.com/settings, /explore, /trending etc.
- Replace setTimeout(120ms) with MutationObserver for robust
  button injection on slow renders
- Add pageshow event for bfcache navigation (back/forward cache)
- Store last 5 mapped repos in chrome.storage.local + "Recent"
  section in popup with one-click reopen
- CHANGELOG v0.4.6 + ROADMAP cleanup (4 done items moved to Done)
