+++ extension/.gitignore
@@ -1,4 +1,6 @@
 node_modules/
 dist/
-package-lock.json
+.output/
+.wxt/
+*.zip
 *.tsbuildinfo
