--- extension/manifest.json
@@ -1,33 +1,0 @@
-{
-  "manifest_version": 3,
-  "name": "StarMapper",
-  "description": "Map the stargazers of any GitHub repository on an interactive world map.",
-  "version": "1.0.0",
-  "action": {
-    "default_popup": "popup.html",
-    "default_icon": {
-      "16": "icons/icon16.png",
-      "48": "icons/icon48.png",
-      "128": "icons/icon128.png"
-    },
-    "default_title": "StarMapper"
-  },
-  "icons": {
-    "16": "icons/icon16.png",
-    "48": "icons/icon48.png",
-    "128": "icons/icon128.png"
-  },
-  "content_scripts": [
-    {
-      "matches": ["https://github.com/*/*"],
-      "js": ["src/content.ts"],
-      "run_at": "document_end"
-    }
-  ],
-  "background": {
-    "service_worker": "src/background.ts",
-    "type": "module"
-  },
-  "permissions": ["tabs", "contextMenus"],
-  "host_permissions": ["https://github.com/*"]
-}
